healthcheck.sh --connect large noise in error log
The --connect test does a tcp connection. There is no
mysql@::1 or mysql@127.0.0.1 user created so the access
is denied, but the healtcheck is correct.

The entries in the error log because of the default
log_warnings=2.

To avoid hitting max_connect_errors (default 100), a non-connection
way is to look at the proc status for listening sockets.

While normally not a direct correlation to a connection, mariadb
listens and is followed by accepting connections immedately without
anything that can fail inbetween.

We move the previous connect to mariadb_connect for compatibility
if users want to continue with that mechanism. We pre-emtively create
tcp listen as a mechanism and redirect connect to that.

Also tested:

lsof -t -p 1 -a -iTCP -sTCP:LISTEN cannot resolve symlink /proc/1/fd/{num}
when run as root.

ss -nlHt state listening "( sport = :3306 )" could work, but no way
to tie to pid 1, and the 3306 port isn't fixed.

Closes #430